refactoring subgraph publishing into operations
